requirements.txt
pyyaml>=6.0
requests>=2.31
pycryptodome>=3.19
